Test Modes
Added testmode.ino which and updated "Test Sensors" module.  Current
test modes are OnOff and TargetforPeriod.  OnOff is standard alarm when
under or over depending on setting for sensor. TargetforPeriod will
alarm until the target has been reached for a set amount of time.  It
also has a "settling" period which can be used to allow time for
over/undershoot to settle out. Set to 0 if you want to skip.